pyproject.toml
[project]
name = "ecommerce-bresilien-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.43.2",
]

[dependency-groups]
dev = [
    "kagglehub>=0.3.10",
]
